chore: transfer ownership of auth & profile sync e2e to identity
