+++ test/e2e/tests/identity/account-syncing/new-user-sync.spec.ts
@@ -2,27 +2,29 @@ import { Mockttp } from 'mockttp';
 import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
 import { withFixtures } from '../../../helpers';
 import FixtureBuilder from '../../../fixture-builder';
-import { mockNotificationServices } from '../mocks';
-import {
-  NOTIFICATIONS_TEAM_IMPORTED_PRIVATE_KEY,
-  NOTIFICATIONS_TEAM_PASSWORD,
-  NOTIFICATIONS_TEAM_SEED_PHRASE,
-} from '../constants';
-import { UserStorageMockttpController } from '../../../helpers/user-storage/userStorageMockttpController';
+import { mockIdentityServices } from '../mocks';
+import { IDENTITY_TEAM_PASSWORD } from '../constants';
+import { UserStorageMockttpController } from '../../../helpers/identity/user-storage/userStorageMockttpController';
 import HeaderNavbar from '../../../page-objects/pages/header-navbar';
 import AccountListPage from '../../../page-objects/pages/account-list-page';
 import HomePage from '../../../page-objects/pages/home/homepage';
-import { completeImportSRPOnboardingFlow } from '../../../page-objects/flows/onboarding.flow';
-import { accountsSyncMockResponse } from './mockData';
+import {
+  completeCreateNewWalletOnboardingFlow,
+  completeImportSRPOnboardingFlow,
+} from '../../../page-objects/flows/onboarding.flow';
+import PrivacySettings from '../../../page-objects/pages/settings/privacy-settings';
+import SettingsPage from '../../../page-objects/pages/settings/settings-page';
 import { IS_ACCOUNT_SYNCING_ENABLED } from './helpers';
 
-describe('Account syncing - Import With Private Key @no-mmi', function () {
+describe('Account syncing - New User @no-mmi', function () {
   if (!IS_ACCOUNT_SYNCING_ENABLED) {
     return;
   }
+
   describe('from inside MetaMask', function () {
-    it('does not sync accounts imported with private keys', async function () {
+    it('syncs after new wallet creation', async function () {
       const userStorageMockttpController = new UserStorageMockttpController();
+      let walletSrp: string;
 
       await withFixtures(
         {
@@ -32,47 +34,60 @@ describe('Account syncing - Import With Private Key @no-mmi', function () {
             userStorageMockttpController.setupPath(
               USER_STORAGE_FEATURE_NAMES.accounts,
               server,
-              {
-                getResponse: accountsSyncMockResponse,
-              },
             );
 
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
           },
         },
         async ({ driver }) => {
-          await completeImportSRPOnboardingFlow({
+          // Create a new wallet
+          await completeCreateNewWalletOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
           await homePage.check_expectedBalanceIsDisplayed();
           await homePage.check_hasAccountSyncingSyncedAtLeastOnce();
 
+          // Open account menu and validate 1 account is shown
           const header = new HeaderNavbar(driver);
           await header.check_pageIsLoaded();
           await header.openAccountMenu();
 
           const accountListPage = new AccountListPage(driver);
           await accountListPage.check_pageIsLoaded();
-          await accountListPage.check_numberOfAvailableAccounts(
-            accountsSyncMockResponse.length,
-          );
+          await accountListPage.check_numberOfAvailableAccounts(1);
           await accountListPage.check_accountDisplayedInAccountList(
-            'My First Synced Account',
+            'Account 1',
           );
-          await accountListPage.check_accountDisplayedInAccountList(
-            'My Second Synced Account',
-          );
+
+          // Add a second account
           await accountListPage.openAccountOptionsMenu();
-          await accountListPage.addNewImportedAccount(
-            NOTIFICATIONS_TEAM_IMPORTED_PRIVATE_KEY,
+          await accountListPage.addNewAccount('My Second Account');
+
+          // Set SRP to use for retreival
+          const headerNavbar = new HeaderNavbar(driver);
+          await headerNavbar.check_pageIsLoaded();
+          await headerNavbar.openSettingsPage();
+          const settingsPage = new SettingsPage(driver);
+          await settingsPage.check_pageIsLoaded();
+          await settingsPage.goToPrivacySettings();
+
+          const privacySettings = new PrivacySettings(driver);
+          await privacySettings.check_pageIsLoaded();
+          await privacySettings.openRevealSrpQuiz();
+          await privacySettings.completeRevealSrpQuiz();
+          await privacySettings.fillPasswordToRevealSrp(
+            IDENTITY_TEAM_PASSWORD,
           );
+          walletSrp = await privacySettings.getSrpInRevealSrpDialog();
+          if (!walletSrp) {
+            throw new Error('Wallet SRP was not set');
+          }
         },
       );
 
@@ -85,35 +100,39 @@ describe('Account syncing - Import With Private Key @no-mmi', function () {
               USER_STORAGE_FEATURE_NAMES.accounts,
               server,
             );
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
           },
         },
         async ({ driver }) => {
+          // Onboard with import flow using SRP from new account created above
           await completeImportSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: walletSrp,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
           await homePage.check_expectedBalanceIsDisplayed();
           await homePage.check_hasAccountSyncingSyncedAtLeastOnce();
 
+          // Open account menu and validate the 2 accounts have been retrieved
           const header = new HeaderNavbar(driver);
           await header.check_pageIsLoaded();
           await header.openAccountMenu();
 
           const accountListPage = new AccountListPage(driver);
           await accountListPage.check_pageIsLoaded();
+
           await accountListPage.check_numberOfAvailableAccounts(2);
+
           await accountListPage.check_accountDisplayedInAccountList(
-            'My First Synced Account',
+            'Account 1',
           );
           await accountListPage.check_accountDisplayedInAccountList(
-            'My Second Synced Account',
+            'My Second Account',
           );
         },
       );
