+++ test/e2e/tests/identity/account-syncing/onboarding-with-opt-out.spec.ts
@@ -2,12 +2,12 @@ import { Mockttp } from 'mockttp';
 import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
 import { withFixtures } from '../../../helpers';
 import FixtureBuilder from '../../../fixture-builder';
-import { mockNotificationServices } from '../mocks';
+import { mockIdentityServices } from '../mocks';
 import {
-  NOTIFICATIONS_TEAM_PASSWORD,
-  NOTIFICATIONS_TEAM_SEED_PHRASE,
+  IDENTITY_TEAM_PASSWORD,
+  IDENTITY_TEAM_SEED_PHRASE,
 } from '../constants';
-import { UserStorageMockttpController } from '../../../helpers/user-storage/userStorageMockttpController';
+import { UserStorageMockttpController } from '../../../helpers/identity/user-storage/userStorageMockttpController';
 import AccountListPage from '../../../page-objects/pages/account-list-page';
 import HeaderNavbar from '../../../page-objects/pages/header-navbar';
 import HomePage from '../../../page-objects/pages/home/homepage';
@@ -45,7 +45,7 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
                 getResponse: accountsSyncMockResponse,
               },
             );
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -54,8 +54,8 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
         async ({ driver }) => {
           await importSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: IDENTITY_TEAM_SEED_PHRASE,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const onboardingCompletePage = new OnboardingCompletePage(driver);
           await onboardingCompletePage.check_pageIsLoaded();
@@ -105,7 +105,7 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
               USER_STORAGE_FEATURE_NAMES.accounts,
               server,
             );
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -114,7 +114,7 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
         async ({ driver }) => {
           await createNewWalletOnboardingFlow({
             driver,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const onboardingCompletePage = new OnboardingCompletePage(driver);
           await onboardingCompletePage.check_pageIsLoaded();
@@ -155,7 +155,7 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
           await privacySettings.openRevealSrpQuiz();
           await privacySettings.completeRevealSrpQuiz();
           await privacySettings.fillPasswordToRevealSrp(
-            NOTIFICATIONS_TEAM_PASSWORD,
+            IDENTITY_TEAM_PASSWORD,
           );
           walletSrp = await privacySettings.getSrpInRevealSrpDialog();
           if (!walletSrp) {
@@ -174,7 +174,7 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
               USER_STORAGE_FEATURE_NAMES.accounts,
               server,
             );
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -184,7 +184,7 @@ describe('Account syncing - Opt-out Profile Sync @no-mmi', function () {
           await completeImportSRPOnboardingFlow({
             driver,
             seedPhrase: walletSrp,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
