+++ test/e2e/tests/identity/account-syncing/sync-after-adding-account.spec.ts
@@ -2,12 +2,12 @@ import { Mockttp } from 'mockttp';
 import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
 import { withFixtures } from '../../../helpers';
 import FixtureBuilder from '../../../fixture-builder';
-import { mockNotificationServices } from '../mocks';
+import { mockIdentityServices } from '../mocks';
 import {
-  NOTIFICATIONS_TEAM_PASSWORD,
-  NOTIFICATIONS_TEAM_SEED_PHRASE,
+  IDENTITY_TEAM_PASSWORD,
+  IDENTITY_TEAM_SEED_PHRASE,
 } from '../constants';
-import { UserStorageMockttpController } from '../../../helpers/user-storage/userStorageMockttpController';
+import { UserStorageMockttpController } from '../../../helpers/identity/user-storage/userStorageMockttpController';
 import HeaderNavbar from '../../../page-objects/pages/header-navbar';
 import AccountListPage from '../../../page-objects/pages/account-list-page';
 import HomePage from '../../../page-objects/pages/home/homepage';
@@ -36,7 +36,7 @@ describe('Account syncing - Add Account @no-mmi', function () {
               },
             );
 
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -45,8 +45,8 @@ describe('Account syncing - Add Account @no-mmi', function () {
         async ({ driver }) => {
           await completeImportSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: IDENTITY_TEAM_SEED_PHRASE,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
@@ -81,7 +81,7 @@ describe('Account syncing - Add Account @no-mmi', function () {
               USER_STORAGE_FEATURE_NAMES.accounts,
               server,
             );
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -90,8 +90,8 @@ describe('Account syncing - Add Account @no-mmi', function () {
         async ({ driver }) => {
           await completeImportSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: IDENTITY_TEAM_SEED_PHRASE,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
@@ -141,7 +141,7 @@ describe('Account syncing - Add Account @no-mmi', function () {
               },
             );
 
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -150,8 +150,8 @@ describe('Account syncing - Add Account @no-mmi', function () {
         async ({ driver }) => {
           await completeImportSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: IDENTITY_TEAM_SEED_PHRASE,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
@@ -186,7 +186,7 @@ describe('Account syncing - Add Account @no-mmi', function () {
               USER_STORAGE_FEATURE_NAMES.accounts,
               server,
             );
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -195,8 +195,8 @@ describe('Account syncing - Add Account @no-mmi', function () {
         async ({ driver }) => {
           await completeImportSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: IDENTITY_TEAM_SEED_PHRASE,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
