+++ test/e2e/tests/identity/account-syncing/sync-after-onboarding.spec.ts
@@ -2,12 +2,12 @@ import { Mockttp } from 'mockttp';
 import { USER_STORAGE_FEATURE_NAMES } from '@metamask/profile-sync-controller/sdk';
 import { withFixtures } from '../../../helpers';
 import FixtureBuilder from '../../../fixture-builder';
-import { mockNotificationServices } from '../mocks';
+import { mockIdentityServices } from '../mocks';
 import {
-  NOTIFICATIONS_TEAM_PASSWORD,
-  NOTIFICATIONS_TEAM_SEED_PHRASE,
+  IDENTITY_TEAM_PASSWORD,
+  IDENTITY_TEAM_SEED_PHRASE,
 } from '../constants';
-import { UserStorageMockttpController } from '../../../helpers/user-storage/userStorageMockttpController';
+import { UserStorageMockttpController } from '../../../helpers/identity/user-storage/userStorageMockttpController';
 import HeaderNavbar from '../../../page-objects/pages/header-navbar';
 import AccountListPage from '../../../page-objects/pages/account-list-page';
 import HomePage from '../../../page-objects/pages/home/homepage';
@@ -15,12 +15,12 @@ import { completeImportSRPOnboardingFlow } from '../../../page-objects/flows/onb
 import { accountsSyncMockResponse } from './mockData';
 import { IS_ACCOUNT_SYNCING_ENABLED } from './helpers';
 
-describe('Account syncing - Rename Accounts @no-mmi', function () {
+describe('Account syncing - Onboarding @no-mmi', function () {
   if (!IS_ACCOUNT_SYNCING_ENABLED) {
     return;
   }
   describe('from inside MetaMask', function () {
-    it('syncs renamed account names', async function () {
+    it('retrieves all previously synced accounts', async function () {
       const userStorageMockttpController = new UserStorageMockttpController();
 
       await withFixtures(
@@ -35,8 +35,7 @@ describe('Account syncing - Rename Accounts @no-mmi', function () {
                 getResponse: accountsSyncMockResponse,
               },
             );
-
-            return mockNotificationServices(
+            return mockIdentityServices(
               server,
               userStorageMockttpController,
             );
@@ -45,8 +44,8 @@ describe('Account syncing - Rename Accounts @no-mmi', function () {
         async ({ driver }) => {
           await completeImportSRPOnboardingFlow({
             driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
+            seedPhrase: IDENTITY_TEAM_SEED_PHRASE,
+            password: IDENTITY_TEAM_PASSWORD,
           });
           const homePage = new HomePage(driver);
           await homePage.check_pageIsLoaded();
@@ -64,55 +63,6 @@ describe('Account syncing - Rename Accounts @no-mmi', function () {
           );
           await accountListPage.check_accountDisplayedInAccountList(
             'My First Synced Account',
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            'My Second Synced Account',
-          );
-          await accountListPage.openAccountOptionsMenu();
-          await accountListPage.changeAccountLabel('My Renamed First Account');
-        },
-      );
-
-      await withFixtures(
-        {
-          fixtures: new FixtureBuilder({ onboarding: true }).build(),
-          title: this.test?.fullTitle(),
-          testSpecificMock: (server: Mockttp) => {
-            userStorageMockttpController.setupPath(
-              USER_STORAGE_FEATURE_NAMES.accounts,
-              server,
-            );
-            return mockNotificationServices(
-              server,
-              userStorageMockttpController,
-            );
-          },
-        },
-        async ({ driver }) => {
-          await completeImportSRPOnboardingFlow({
-            driver,
-            seedPhrase: NOTIFICATIONS_TEAM_SEED_PHRASE,
-            password: NOTIFICATIONS_TEAM_PASSWORD,
-          });
-          const homePage = new HomePage(driver);
-          await homePage.check_pageIsLoaded();
-          await homePage.check_expectedBalanceIsDisplayed();
-          await homePage.check_hasAccountSyncingSyncedAtLeastOnce();
-
-          const header = new HeaderNavbar(driver);
-          await header.check_pageIsLoaded();
-          await header.openAccountMenu();
-
-          const accountListPage = new AccountListPage(driver);
-          await accountListPage.check_pageIsLoaded();
-          await accountListPage.check_numberOfAvailableAccounts(
-            accountsSyncMockResponse.length,
-          );
-          await accountListPage.check_accountIsNotDisplayedInAccountList(
-            'My First Synced Account',
-          );
-          await accountListPage.check_accountDisplayedInAccountList(
-            'My Renamed First Account',
           );
           await accountListPage.check_accountDisplayedInAccountList(
             'My Second Synced Account',
