+++ test/e2e/tests/identity/constants.ts
@@ -1,0 +1,9 @@
+// As we rely on profile syncing for most of our features, we need to use the same SRP for all of our tests
+export const IDENTITY_TEAM_SEED_PHRASE =
+  'leisure swallow trip elbow prison wait rely keep supply hole general mountain';
+export const IDENTITY_TEAM_PASSWORD = 'notify_password';
+// You can use the storage key below to generate mock data
+export const IDENTITY_TEAM_STORAGE_KEY =
+  '0d55d30da233959674d14076737198c05ae3fb8631a17e20d3c28c60dddd82f7';
+export const IDENTITY_TEAM_IMPORTED_PRIVATE_KEY =
+  '0179f7453ff337aba89d04b00085764092cf8c0cfe91d5614c39c2be902ad582';
